backend/requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
Flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
nltk==3.9.2
packaging==25.0
pypdf==6.1.1
python-dotenv==1.1.1
regex==2025.9.18
requests==2.32.5
tqdm==4.67.1
urllib3==2.5.0
Werkzeug==3.1.3
